Add Subscriber
Pasted in the original code for login.html and login.html,adjusted the save controller in SubscriberController and layouttemplate.html. layouttemplate.html doesn't include the authentication code that Love wrote but can be added later. Non users can subscribe to the newsletter and the field cannot be
empty when you click the subscribe button.